xUnit test basic

+++ SoftPlan.Api.InterestCalculator/Repositories/IInterestRepository.cs
@@ -4,7 +4,7 @@ namespace SoftPlan.Api.InterestCalculator.Repositories
 {
     public interface IInterestRepository
     {
-        double GetInterestRate(string urlApi);
+        decimal GetInterestRate(string urlApi);
         decimal Calculator(decimal initialValue, int month, string urlApi );
     }
 }
